requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
awscli==1.14.44
boto3==1.9.237
botocore==1.12.237
Brlapi==0.6.6
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
colorama==0.3.7
command-not-found==0.3
cryptography==2.7
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==2.2
django-braces==1.13.0
django-cognito-jwt==0.0.3
django-crispy-forms==1.7.2
django-extensions==2.2.1
django-filter==2.2.0
django-s3-storage==0.12.5
django-warrant==0.1.1
djangorestframework==3.10.3
docutils==0.15.2
ecdsa==0.13.2
envs==1.3
future==0.17.1
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-automl==0.5.0
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
httplib2==0.13.1
idna==2.8
jmespath==0.9.4
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
pipenv==2018.11.26
protobuf==3.9.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.3.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.7.1
PyLin==0.2.7
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
pyserial==3.4
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
python-jose-cryptodome==1.3.2
pytz==2019.2
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
roman==2.0.0
rsa==4.0
s3transfer==0.2.1
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.25.6
usb-creator==0.3.3
virtualenv==16.7.5
virtualenv-clone==0.5.3
wadllib==1.3.2
warrant==0.6.1
xkit==0.0.0
zope.interface==4.3.2
